requirements.txt
Flask
Flask-Cors
huggingface-hub
ipython
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llamaindex-py-client
nltk
numpy
openai
packaging
pandas
pandocfilters
paramiko
pydantic
pydantic_core
pymilvus
pypdf
python-dotenv
ragas
requests
sentence-transformers
SQLAlchemy
tqdm
transformers
urllib3
waitress